Update README.md
Overview
Project of "Data Management and Data Visualization" courses, consists in answer the following question: "Is there a better time or airline to travel? Are there any differences between years?" 
We used to the data sources from the website of the Department of Transportation of the United States of America and in particular from the government agency "Bureau of Transportation Statistics" (BTS), it was possible to download data relating to domestic US flights for the years 2018 and 2019, which were then supplemented with information from other sources. With regard to the latter, on the Kaggle platform it was possible to download datasets containing additional information regarding airlines and airports. Finally, from the website of the “Airports Council International - North America” (ACI) data relating to the number of passengers for some of the main airports over several years were obtained. Also, thanks to Python language and MongoDBFor example, the data on the altitude and on the full name of the airports were integrated using the unique code of the airport obtained from the BTS website. 
Finally, the second part of the work shows some of the infographics, created using the Tableau Desktop software.